requirements.txt
librosa
pandas
sed_scores_eval
timm
nnAudio
Cython
av>=10.0.0
h5py>=3.8.0
jsonpickle>=3.0.1
hf_transfer>=0.1.4
hf-fastup>=0.0.5
datasets>=2.15.0
pytorch-lightning>=2.0.0
wandb
transformers
sed_scores_eval==0.0.3
intervaltree
more-itertools